+++ stream-access-control/access-control-core/src/main/java/com/backbase/stream/service/UserService.java
@@ -15,6 +15,9 @@ import com.backbase.dbs.user.api.service.v2.model.Realm;
 import com.backbase.dbs.user.api.service.v2.model.UpdateIdentityRequest;
 import com.backbase.dbs.user.api.service.v2.model.UserCreated;
 import com.backbase.dbs.user.api.service.v2.model.UserExternal;
+import com.backbase.identity.integration.api.service.v1.IdentityIntegrationServiceApi;
+import com.backbase.identity.integration.api.service.v1.model.EnhancedUserRepresentation;
+import com.backbase.identity.integration.api.service.v1.model.UserRequestBody;
 import com.backbase.stream.legalentity.model.IdentityUserLinkStrategy;
 import com.backbase.stream.legalentity.model.LegalEntity;
 import com.backbase.stream.legalentity.model.User;
@@ -23,6 +26,7 @@ import com.backbase.stream.mapper.UserMapper;
 import java.text.MessageFormat;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
@@ -46,6 +50,7 @@ public class UserService {
 
     private final UserManagementApi usersApi;
     private final IdentityManagementApi identityManagementApi;
+    private final Optional<IdentityIntegrationServiceApi> identityIntegrationApi;
 
     /**
      * Get User by external ID.
@@ -225,27 +230,55 @@ public class UserService {
                 user.setExternalId(identityCreatedItem.getExternalId());
                 return user;
             })
-            .filter(u -> IdentityUserLinkStrategy.IMPORT_FROM_IDENTIY.equals(u.getIdentityLinkStrategy()))  //if not IMPORT_FROM_IDENTIY this will return Mono.empty()! this means for instance that user internalId will not be returned although user was created
             .flatMap(this::updateIdentityUserAttributes);
     }
 
+    private Mono<User> updateIdentityUserAttributes(User user) {
+        if (IdentityUserLinkStrategy.IMPORT_FROM_IDENTIY.equals(user.getIdentityLinkStrategy())
+            && user.getAttributes() != null) {
+            UpdateIdentityRequest replaceIdentity = new UpdateIdentityRequest();
+            replaceIdentity.attributes(user.getAttributes());
+            return identityManagementApi.updateIdentity(user.getInternalId(), replaceIdentity)
+                .doOnError(WebClientResponseException.BadRequest.class, badRequest ->
+                    log.error("Error adding user attributes: {}", badRequest.getResponseBodyAsString()))
+                .then(Mono.just(user));
+        }
+        return Mono.just(user);
+    }
 
     /**
-     * Update identity user attributes
-     *
-     * @param user
-     * @return {@link Mono<Void>}
+     * Locks/Unlocks the user is current status is different.
+     * @param user user to be locked/unlocked.
+     * @param realm user's realm.
+     * @return user.
      */
-    public Mono<User> updateIdentityUserAttributes(User user) {
-        if (user.getAttributes() == null)
-            Mono.just(user);
+    public Mono<User> changeEnableStatus(User user, String realm) {
+        return identityIntegrationApi.map(api -> getIdentityUser(user, realm)
+            .flatMap(eur -> {
+                boolean shouldEnable = user.getLocked() != null ? !user.getLocked() : eur.getEnabled();
+                if (!eur.getEnabled().equals(shouldEnable)) {
+                    UserRequestBody presentationUser = mapper.toPresentation(eur);
+                    presentationUser.setEnabled(shouldEnable);
+                    return api.updateUserById(realm, user.getInternalId(), presentationUser);
+                }
+                return Mono.just(user);
+            })
+            .doOnNext(eur -> {
+                log.info("User {} locked successfully", user.getExternalId());
+            })
+            .doOnError(WebClientResponseException.class, e ->
+                log.error("Error locking user {}: {}", user.getInternalId(), e.getResponseBodyAsString()))
+            .thenReturn(user)).orElse(Mono.just(user));
+    }
 
-        UpdateIdentityRequest replaceIdentity = new UpdateIdentityRequest();
-        replaceIdentity.attributes(user.getAttributes());
-        return identityManagementApi.updateIdentity(user.getInternalId(), replaceIdentity)
-            .doOnError(WebClientResponseException.BadRequest.class, badRequest ->
-                log.error("Error adding user attributes: {}", badRequest.getResponseBodyAsString()))
-            .then(Mono.just(user));
+    private Mono<EnhancedUserRepresentation> getIdentityUser(User user, String realm) {
+        return identityIntegrationApi.map(api -> api.getUserById(realm, user.getInternalId())
+            .doOnNext(eur -> {
+                log.info("Identity user found: {}", user.getExternalId());
+            })
+            .doOnError(WebClientResponseException.class, e ->
+                log.error("Error retrieving identity user {}: {}", user.getInternalId(), e.getResponseBodyAsString())))
+            .orElse(Mono.empty());
     }
 
     private User handleCreateUserResult(User user, UserCreated userCreated) {